+++ src/python/dxpy/bindings/dxproject.py
@@ -367,7 +367,7 @@ class DXProject(DXContainer):
                allowed_executables=None, unset_allowed_executables=None,
                database_ui_view_only=None, external_upload_restricted=None,
                database_results_restricted=None, unset_database_results_restricted=None,
-               https_app_isolated_browsing=None, **kwargs):
+               https_app_isolated_browsing=None, https_app_isolated_browsing_options=None, **kwargs):
         """
         :param name: If provided, the new project name
         :type name: string
@@ -398,6 +398,8 @@ class DXProject(DXContainer):
         :param https_app_isolated_browsing: Whether all https access to jobs in this project should be wrapped in Isolated Browsing.
                                             If true, httpsApp-enabled executables must have httpsApp.shared_access set to 'NONE' to run in this project.
         :type https_app_isolated_browsing: boolean
+        :param https_app_isolated_browsing_options: Options for Isolated Browsing. See https://documentation.dnanexus.com/developer/api/data-containers/projects#api-method-project-xxxx-update for a list of supported keys.
+        :type https_app_isolated_browsing_options: dict
 
         Updates the project with the new fields. All fields are
         optional. Fields that are not provided are not changed.
@@ -435,6 +437,8 @@ class DXProject(DXContainer):
             update_hash["databaseResultsRestricted"] = None
         if https_app_isolated_browsing is not None:
             update_hash["httpsAppIsolatedBrowsing"] = https_app_isolated_browsing
+        if https_app_isolated_browsing_options is not None:
+            update_hash["httpsAppIsolatedBrowsingOptions"] = https_app_isolated_browsing_options
         dxpy.api.project_update(self._dxid, update_hash, **kwargs)
 
     def invite(self, invitee, level, send_email=True, **kwargs):